pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shellgpt"
version = "1.0.0"
description = "🤖 AI-powered intelligent shell assistant that understands natural language"
authors = [
    {name = "ychenfen", email = "2570601904@qq.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["ai", "shell", "cli", "assistant", "natural-language", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Shells",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "psutil>=5.9.0",
    "prompt-toolkit>=3.0.0",
    "typer>=0.9.0",
    "requests>=2.28.0",
    "pyyaml>=6.0.0",
    "platformdirs>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
homepage = "https://github.com/ychenfen/shellgpt"
repository = "https://github.com/ychenfen/shellgpt"
documentation = "https://github.com/ychenfen/shellgpt#readme"
changelog = "https://github.com/ychenfen/shellgpt/blob/main/CHANGELOG.md"

[project.scripts]
shellgpt = "shellgpt.cli.main:app"
sgpt = "shellgpt.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["shellgpt*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=shellgpt --cov-report=html --cov-report=term-missing"